mediacapture-streams/MediaStream-removetrack.https.html

<!doctype html>
<html>
<head>
<title>Removing a track from a MediaStream</title>
<link rel="author" title="Dominique Hazael-Massieux" href="mailto:dom@w3.org"/>
<link rel="help" href="http://dev.w3.org/2011/webrtc/editor/getusermedia.html#widl-MediaStreamTrackList-remove-void-MediaStreamTrack-track">
<link rel="help" href="http://dev.w3.org/2011/webrtc/editor/getusermedia.html#event-mediastream-removetrack">
</head>
<body>
<p class="instructions">When prompted, accept to share your audio stream, then your video stream.</p>
<h1 class="instructions">Description</h1>
<p class="instructions">This test checks that removinging a track from a MediaStream works as expected.</p>
<div id='log'></div>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>

promise_test(async t => {
  const stream = await navigator.mediaDevices.getUserMedia({video: true, audio: true});
  const tracks = stream.getTracks();
  t.add_cleanup(() => tracks.forEach(track => track.stop()));
  const stream2 = await navigator.mediaDevices.getUserMedia({audio: true});
  tracks.push(...stream2.getTracks());

  stream.onremovetrack = stream2.onremovetrack = t.step_func(() =>
    assert_unreached("onremovetrack is not triggered by script itself"));

  assert_equals(stream.getTracks().length, 2, "mediastream starts with 2 tracks");
  stream.removeTrack(stream.getVideoTracks()[0]);
  assert_equals(stream.getTracks().length, 1, "mediastream has 1 track left");
  stream.removeTrack(stream.getAudioTracks()[0]);
  assert_equals(stream.getTracks().length, 0, "mediastream has no tracks left");
  stream.removeTrack(stream2.getTracks()[0]); // should not throw

  // Allow time to verify no events fire.
  await new Promise(r => t.step_timeout(r, 1));

}, "Tests that a removal from a MediaStream works as expected");

</script>
</body>
</html>
